Add subtitle keyword input for filtering subtitle tracks.
This will allow an optional keyword to be supplied. This keyword needs to be present in either the title or displayTitle in order to be considered for matching.

I made this change so I can easily make subtitles with "Forced" in the title be the only subtitles applied. This can be improved if there is a way to determine the forced flag instead of it needing to be in the name, but for my library this was sufficient and helpful.